Restructure routes, and add icon to dashboard/forms page
