+++ app/client/src/components/App.tsx
@@ -11,11 +11,7 @@ function App() {
 
       <div className="display-flex flex-justify border-bottom padding-bottom-1">
         <nav>
-          <Link to="/rebates">Rebates</Link>
-        </nav>
-
-        <nav>
-          <Link to="/profile" className="usa-button">
+          <Link to="/" className="usa-button usa-button--outline">
             <span className="display-flex flex-align-center">
               <svg
                 className="usa-icon margin-right-1"
@@ -23,7 +19,26 @@ function App() {
                 focusable="false"
                 role="img"
               >
-                <use href={`${icons}#account_circle`}></use>
+                <use href={`${icons}#list`} />
+              </svg>
+              Forms
+            </span>
+          </Link>
+        </nav>
+
+        <nav>
+          <Link
+            to="/profile"
+            className="usa-button usa-button--outline margin-right-0"
+          >
+            <span className="display-flex flex-align-center">
+              <svg
+                className="usa-icon margin-right-1"
+                aria-hidden="true"
+                focusable="false"
+                role="img"
+              >
+                <use href={`${icons}#account_circle`} />
               </svg>
               Profile
             </span>
